[DEV] Refactor/code (#52)
* refactor(auth): Improve splitting code in auth directory

* refactor(dashboard): Splitting dashboard code and fix code structure

* refactor(chatbot): Split code in chatbot

* refactor(dashboard): Code splitting and refactor structure

* refactor(saved): Code splitting and refactor code structure

* refactor(auth):  Split code and refactor folder structure

* feat(utils): Separate the alert from the component

* feat(service): Add service layer for profile

* feat(error): Add error page for profile

* feat(data): Add data for profile navigation link

* feat(props): Separate props for profile and it's component

* feat(loading): Add loading page for profile route

* feat(custom-field): Add custom field input component based on it's variant

* feat(route): Add route for profile page

* feat(profile-content): Add profile component page

* feat(field): Add custom input field for profile

* feat(component): Create profile form component

* feat(form field): Add custom select field component

* refactor(route): Move route for favorite component

* refactor(profile): Split code for better maintenance and readability

* chore: Update eslintrc json

* chore: Update support component
