+++ .eslintrc.json
@@ -49,7 +49,7 @@
     "no-unsafe-negation": "error",
     "valid-typeof": ["error", { "requireStringLiterals": true }],
     "linebreak-style": ["error", "windows"],
-    "semi":"off",
+    "semi": "off",
 
     // Best Practices
     "curly": "error",
@@ -117,7 +117,5 @@
 
     // typescript
     "@typescript-eslint/no-explicit-any": "off"
-  
-    
   }
 }
