+++ app/(auth)/index.props.ts
@@ -1,0 +1,5 @@
+import React from 'react'
+
+export type AuthProps = {
+  children: React.ReactNode
+}
