+++ app/(route)/dashboard/components/LogoIcon.tsx
@@ -1,10 +1,6 @@
-import { NextPage } from 'next'
 import Link from 'next/link'
-import { motion } from 'framer-motion'
 
-interface Props {}
-
-const Logo: NextPage<Props> = () => {
+const LogoIcon = () => {
   return (
     <>
       <Link
@@ -13,16 +9,9 @@ const Logo: NextPage<Props> = () => {
         className='relative z-20 flex items-center space-x-2 py-1 text-sm font-normal text-black'
       >
         <div className='h-5 w-6 flex-shrink-0 rounded-bl-sm rounded-br-lg rounded-tl-lg rounded-tr-sm bg-cyan-800 dark:bg-white' />
-        <motion.span
-          initial={{ opacity: 0 }}
-          animate={{ opacity: 1 }}
-          className='whitespace-pre font-medium text-black dark:text-white'
-        >
-          Melancong
-        </motion.span>
       </Link>
     </>
   )
 }
 
-export default Logo
+export default LogoIcon
