+++ app/(route)/dashboard/profile/components/ProfileContent.tsx
@@ -1,0 +1,12 @@
+import React from 'react'
+import ProfileForm from './form/ProfileForm'
+
+const ProfileContent = () => {
+  return (
+    <>
+      <ProfileForm />
+    </>
+  )
+}
+
+export default ProfileContent
