+++ app/(route)/dashboard/profile/index.data.tsx
@@ -1,0 +1,15 @@
+import {
+  IconBrandHipchat,
+  IconLockSquare,
+  IconSettings
+} from '@tabler/icons-react'
+
+export const supportItems = [
+  { text: 'Terms and Conditions', icon: <IconSettings />, link: '#' },
+  { text: 'Privacy Policy', icon: <IconLockSquare />, link: '#' },
+  {
+    text: 'Need Help? Let`s chat',
+    icon: <IconBrandHipchat />,
+    link: 'https://wa.me/62881037217902?text=Hello Admin, i`m looking for some help regarding your services. Could you please assist me?'
+  }
+]
