+++ app/(route)/dashboard/profile/loading.tsx
@@ -1,0 +1,11 @@
+const ProfileLoading = () => {
+  return (
+    <div className='mx-auto size-full min-h-dvh rounded-lg bg-white p-5'>
+      <div className='flex size-full min-h-dvh items-center justify-center'>
+        <div className='h-10 w-10 animate-spin rounded-full border-4 border-gray-200 border-t-blue-500'></div>
+      </div>
+    </div>
+  )
+}
+
+export default ProfileLoading
