+++ app/(route)/dashboard/saved/index.props.ts
@@ -1,0 +1,13 @@
+export type ErrorResponseSavedTypes = {
+  response: {
+    status: number
+  }
+}
+export type DestinationSavedTypes = {
+  id: string
+  name: string
+  rating: string
+  regency: string
+  imageLink: string
+  information: string
+}
